+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.order.dto.OrderTableRequest;
 
 public class TableGroupRequest {
