+++ src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.table.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.table.application.TableService;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
